+++ test_dcm_import_module/conftest.py
@@ -23,6 +23,18 @@ def _fixtures():
     return path
 
 
+@pytest.fixture(scope="session", autouse=True)
+def disable_extension_logging():
+    """
+    Disables the stderr-logging via the helper method `print_status`
+    of the `dcm_common.services.extensions`-subpackage.
+    """
+    # pylint: disable=import-outside-toplevel
+    from dcm_common.services.extensions.common import PrintStatusSettings
+
+    PrintStatusSettings.silent = True
+
+
 @pytest.fixture(name="testing_config")
 def _testing_config(file_storage):
     """Returns test-config"""
@@ -46,7 +58,7 @@ def _client(testing_config):
     Returns test_client.
     """
 
-    return app_factory(testing_config()).test_client()
+    return app_factory(testing_config(), block=True).test_client()
 
 
 @pytest.fixture(name="minimal_request_body_external")