+++ test_dcm_import_module/test_sdk.py
@@ -43,7 +43,7 @@ def test_default_ping(
 ):
     """Test default endpoint `/ping-GET`."""
 
-    run_service(app)
+    run_service(app, probing_path="ready")
 
     response = default_sdk.ping()
 
@@ -55,7 +55,7 @@ def test_default_status(
 ):
     """Test default endpoint `/status-GET`."""
 
-    run_service(app)
+    run_service(app, probing_path="ready")
 
     response = default_sdk.get_status()
 
@@ -68,7 +68,7 @@ def test_default_identify(
 ):
     """Test default endpoint `/identify-GET`."""
 
-    run_service(app)
+    run_service(app, probing_path="ready")
 
     response = default_sdk.identify()
 
@@ -94,7 +94,7 @@ def test_import_report(
 ):
     """Test endpoints `/import/external-POST` and `/report-GET`."""
 
-    run_service(app)
+    run_service(app, probing_path="ready")
     run_service(
         app=fake_builder_service,
         port=8083
@@ -132,7 +132,7 @@ def test_import_report_404(
 ):
     """Test build endpoint `/report-GET` without previous submission."""
 
-    run_service(app)
+    run_service(app, probing_path="ready")
 
     with pytest.raises(dcm_import_module_sdk.rest.ApiException) as exc_info:
         import_sdk.get_report(token="some-token")